pyproject.toml
[project]
name = "Discretized-Optimized-Map-Projection"
version = "0.0.1"
authors = [
  { name="Franz-Benjamin Mocnik", email="mail@mocnik-science.net" },
]
description = "Discretized Optimized Map Projection"
readme = "README.md"
keywords = ["projection", "cartography"]
license = {text = ""}
requires-python = ">=3.9"
classifiers = [
  "Programming Language :: Python :: 3"
]
dependencies = [
  "altair>=5.2.0",
  "multiprocess>=0.70.16",
  "pillow>=10.2.0",
  "py>=1.11.0",
  "pyproj>=3.6.1",
  "pyshp>=2.3.1",
  "requests>=2.31.0",
  "scikit-learn>=1.4.0",
  "scipy>=1.12.0",
  "shapely>=2.0.2",
  "drawsvg~=2.0",
  "vl-convert-python>=1.2.3",
  "wxpython>=4.2.1",
]

[project.urls]
Homepage = "https://github.com/mocnik-science/..."
